requirements.txt
fastapi
uvicorn
pydantic
requests
pdfplumber
python-multipart
python-dotenv
